Cleaning Centos container
The actual container is running a pretty classic ubuntu container full of useless stuff.
As a result the container is pretty big on the filesystem.

This patch aims at gaining disk's space by using the following techniques:
- stripping raw binaries
- removing useless data files like documentation
- removing useless binaries
- removing useless packages like perl
- removing testing code/binaries
- removing pre-compiled python files as we nearly don't use them
- compressing debugging tools
- compressing fat binaries without loosing it executability
- clearing log files
- moving all the disk's consuming tasks in a single RUN :
Cleaning in a last layer makes the filesystem smaller but the whole
image is still big as former layers are still big.
Installing the ceph packages & confd/kubectl is clearly the biggest task.
Running the clean process in the same RUN as the yum install &
wget makes this layer thinner and saves ~400MB of storage.

Note the stripping process is done in two passes :
- first one targets "easy" stuff
- second one targets more tricky such that may be discussed if some
issue occurs

A typical output looks like the following :
	Stripping process: Step 1 removed 119MB
	[...]
	Stripping process: Step 2 removed 89MB
	[...]
	Stripping process saved 208MB and dropped container size from 705MB to 497MB